+++ library/newlib/close.c
@@ -1,7 +1,7 @@
 #include <svc.h>
 #include <portal.h>
 
-int close(unsigned int file)
+int close(long file)
 {
     return (int)portal_request(PORTAL_VFS, 0, (void *)file, 1, 0);
 }
